+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes
 
 0.8 (Unreleased)
 ----------------
+ - launchpad #1130838: build-only-if-requires buildslave option
  - Using the uniform test launcher script provided by anybox.recipe.openerp 1.2
  - launchpad #1086066: detecting unittest2 failures and errors
  - launchpad #1086392: resilience wrt missing remote mercurial